restructured triggers files
