+++ dlp/snippets/Triggers/triggers_test.py
@@ -18,14 +18,19 @@ import uuid
 
 import google.api_core.exceptions
 import google.cloud.storage
+
+import create_trigger as ct
+import delete_trigger as dt
+import list_triggers as lt
+
 import pytest
 
-import triggers
+import update_trigger as ut
 
 UNIQUE_STRING = str(uuid.uuid4()).split("-")[0]
 GCLOUD_PROJECT = os.getenv("GOOGLE_CLOUD_PROJECT")
 TEST_BUCKET_NAME = GCLOUD_PROJECT + "-dlp-python-client-test" + UNIQUE_STRING
-RESOURCE_DIRECTORY = os.path.join(os.path.dirname(__file__), "resources")
+RESOURCE_DIRECTORY = os.path.join(os.path.dirname(__file__), "../resources")
 RESOURCE_FILE_NAMES = ["test.txt", "test.png", "harmless.txt", "accounts.txt"]
 TEST_TRIGGER_ID = "test-trigger" + UNIQUE_STRING
 
@@ -67,7 +72,7 @@ def test_create_list_update_and_delete_trigger(
     bucket: google.cloud.storage.bucket.Bucket, capsys: pytest.CaptureFixture
 ) -> None:
     try:
-        triggers.create_trigger(
+        ct.create_trigger(
             GCLOUD_PROJECT,
             bucket.name,
             7,
@@ -76,13 +81,13 @@ def test_create_list_update_and_delete_trigger(
         )
     except google.api_core.exceptions.InvalidArgument:
         # Trigger already exists, perhaps due to a previous interrupted test.
-        triggers.delete_trigger(GCLOUD_PROJECT, TEST_TRIGGER_ID)
+        dt.delete_trigger(GCLOUD_PROJECT, TEST_TRIGGER_ID)
 
         out, _ = capsys.readouterr()
         assert TEST_TRIGGER_ID in out
 
         # Try again and move on.
-        triggers.create_trigger(
+        ct.create_trigger(
             GCLOUD_PROJECT,
             bucket.name,
             7,
@@ -94,12 +99,12 @@ def test_create_list_update_and_delete_trigger(
     out, _ = capsys.readouterr()
     assert TEST_TRIGGER_ID in out
 
-    triggers.list_triggers(GCLOUD_PROJECT)
+    lt.list_triggers(GCLOUD_PROJECT)
 
     out, _ = capsys.readouterr()
     assert TEST_TRIGGER_ID in out
 
-    triggers.update_trigger(
+    ut.update_trigger(
         GCLOUD_PROJECT,
         ["US_INDIVIDUAL_TAXPAYER_IDENTIFICATION_NUMBER"],
         TEST_TRIGGER_ID,
@@ -109,7 +114,7 @@ def test_create_list_update_and_delete_trigger(
     assert TEST_TRIGGER_ID in out
     assert "US_INDIVIDUAL_TAXPAYER_IDENTIFICATION_NUMBER" in out
 
-    triggers.delete_trigger(GCLOUD_PROJECT, TEST_TRIGGER_ID)
+    dt.delete_trigger(GCLOUD_PROJECT, TEST_TRIGGER_ID)
 
     out, _ = capsys.readouterr()
     assert TEST_TRIGGER_ID in out
